fix(esp_lvgl_port): Build PPA/DMA2D draw units against registry LVGL
Add an lvgl_public.h shim, include the mask rect descriptor definition,
and widen set_pending_trans to uint16_t so the units compile with LVGL
from the Component Registry (validated with 9.5.0).

+++ components/esp_lvgl_port/src/lvgl9/draw/ppa/lv_draw_ppa.h
@@ -120,7 +120,7 @@ esp_err_t lv_draw_ppa_set_burst_length(lv_draw_ppa_client_kind_t client,
  * @return ESP_OK on success.
  */
 esp_err_t lv_draw_ppa_set_pending_trans(lv_draw_ppa_client_kind_t client,
-                                        uint8_t pending);
+                                        uint16_t pending);
 #endif
 
 #if LV_USE_PPA_STATS
